HYPR-351 Release setup build envs
Build environments for release setup are
mostly complete now. There are probably
some fixes to be done here and there,
but all guest VM components should build
now properly.

README hasn't been updated yet, and building
the full CAmkES system hasn't been tried yet.

+++ scripts/build_sel4.sh
@@ -1,17 +1,67 @@
-#! /bin/sh
+#!/bin/sh
 
 set -e
 
-. `pwd`/.config
+SCRIPT_NAME=$(realpath $0)
+SCRIPT_DIR=$(dirname ${SCRIPT_NAME})
 
-if test "x`pwd`" != "x/workspace"; then
-  exec docker/enter_container.sh `pwd` sel4_build scripts/build_sel4.sh $@
+. "${SCRIPT_DIR}/utils.sh"
+. "$(pwd)/.config"
+
+if test "x$(pwd)" != "x/workspace"; then
+  exec docker/enter_container.sh -i sel4 -w $(pwd) scripts/build_sel4.sh $@
 fi
 
-BUILDDIR=$1
-shift
+BUILDDIR=""
+SRCDIR=""
+OTHER_ARGS=""
 
-SRCDIR=$1
+SKIP=false
+# Parse arguments
+for ARG in "$@"; do
+  if $SKIP; then
+    SKIP=false && continue
+  fi
+  case "$ARG" in
+    -b)
+      shift
+      if test -n "$BUILDDIR"; then
+        die "ERROR: Multiple definitions for parameter BUILDDIR (-b)!"
+      else
+        BUILDDIR="$1"
+      fi
+      SKIP=true
+      ;;
+    -s)
+      shift
+      if test -n "$SRCDIR"; then
+        die "ERROR: Multiple definitions for parameter SRCDIR (-s)!"
+      else
+        SRCDIR="$1"
+      fi
+      SKIP=true
+      ;;
+    *)
+      # Treat everything else as args to init-build
+      OTHER_ARGS="$OTHER_ARGS $ARG "
+      ;;
+  esac
+  shift
+done
+
+# Validate input arguments.
+# Build and source directories
+# are required.
+#
+if test -z "$BUILDDIR"; then
+	die "ERROR: Parameter BUILDDIR (-b) is required!"
+fi
+
+if test -z "$SRCDIR"; then
+	die "ERROR: Parameter SRCDIR (-s) is required!"
+fi
+
+set -- "$OTHER_ARGS"
 
 rm -rf ${BUILDDIR}
 mkdir -p ${BUILDDIR}
@@ -21,5 +71,9 @@ cd ${BUILDDIR}
 ./init-build.sh -B . -DAARCH64=1 -DPLATFORM=${PLATFORM} -DCROSS_COMPILER_PREFIX=${CROSS_COMPILE} $@
 ninja
 
+echo "----------------------------------------------"
+echo "----------------------------------------------"
+echo "----------------------------------------------"
+echo "                                              "
 echo "Here are your binaries in ${BUILDDIR}/images: "
-ls -l ./images
+ls -la ./images
